+++ src/main/java/pokeraidbot/domain/emote/Emotes.java
@@ -15,4 +15,5 @@ public class Emotes {
     public static final String SAD = "\uD83D\uDC94";
     public static final String HAPPY = "\uD83D\uDE03";
     public static final String GROUP = "\uD83D\uDC65";
+    public static final String STAR = "⭐";
 }