+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.8.0 (?)
+=====
+* Possible to mark EX gyms if one includes a gyms_REGION.csv.ex.txt file with a list of the possible EX gyms
+* Mew and Celebi prepared to be raid bosses (max CP etc)
+
 1.7.14 (2018-01-30)
 =====
 * Gym removed from Uppsala